Add scaladocs and move a couple things out of arrivals-api.
