+++ arrivals-api/src/main/scala/com/pagerduty/arrivals/api/aggregator/OneStepAggregator.scala
@@ -3,6 +3,12 @@ package com.pagerduty.arrivals.api.aggregator
 import akka.NotUsed
 import akka.http.scaladsl.model.HttpResponse
 
+/** An [[Aggregator]] that builds a response from a single set of upstream requests, and does not need to accumulate state.
+  *
+  * @tparam AuthData The type of authentication data that this aggregator will provide
+  * @tparam RequestKey The type of key to use as a unique identification for generated requests
+  * @tparam AddressingConfig The type of configuration needed to proxy requests upstream
+  */
 trait OneStepAggregator[AuthData, RequestKey, AddressingConfig]
     extends Aggregator[AuthData, RequestKey, NotUsed, AddressingConfig] {
 
@@ -12,5 +18,6 @@ trait OneStepAggregator[AuthData, RequestKey, AddressingConfig]
     buildOutgoingResponseStateless(upstreamResponses)
   }
 
+  /** Build an outgoing response based on the responses to all upstream requests. */
   def buildOutgoingResponseStateless(upstreamResponses: ResponseMap): HttpResponse
 }
