+++ arrivals-api/src/main/scala/com/pagerduty/arrivals/api/auth/AuthFailedReason.scala
@@ -1,3 +1,8 @@
 package com.pagerduty.arrivals.api.auth
 
-class AuthFailedReason(val metricTag: String)
+/** A reason authentication (as defined in [[AuthenticationConfig.authenticate]]) failed. */
+trait AuthFailedReason {
+
+  /** Define the tag which will be appended to the authentication failure metric. */
+  def metricTag: String
+}
