+++ arrivals/src/main/scala/com/pagerduty/arrivals/impl/aggregator/Aggregator.scala
@@ -5,7 +5,7 @@ import akka.http.scaladsl.unmarshalling.Unmarshal
 import akka.stream.Materializer
 import com.pagerduty.arrivals.api
 import com.pagerduty.arrivals.api.headerauth.HeaderAuthConfig
-import com.pagerduty.arrivals.api.proxy.HttpProxy
+import com.pagerduty.arrivals.impl.proxy.HttpProxy
 
 import scala.concurrent.{ExecutionContext, Future}
