+++ arrivals/src/main/scala/com/pagerduty/arrivals/impl/aggregator/AggregatorDependencies.scala
@@ -2,7 +2,7 @@ package com.pagerduty.arrivals.impl.aggregator
 
 import akka.stream.Materializer
 import com.pagerduty.arrivals.api.headerauth.HeaderAuthConfig
-import com.pagerduty.arrivals.api.proxy.HttpProxy
+import com.pagerduty.arrivals.impl.proxy.HttpProxy
 
 import scala.concurrent.ExecutionContext
