+++ arrivals/src/main/scala/com/pagerduty/arrivals/impl/proxy/HttpProxyLike.scala
@@ -1,0 +1,10 @@
+package com.pagerduty.arrivals.impl.proxy
+
+import akka.http.scaladsl.model.{HttpRequest, HttpResponse}
+import com.pagerduty.arrivals.api.proxy.Upstream
+
+import scala.concurrent.Future
+
+trait HttpProxyLike[AddressingConfig] {
+  def apply(request: HttpRequest, upstream: Upstream[AddressingConfig]): Future[HttpResponse]
+}
