+++ version.sbt
@@ -1,1 +1,1 @@
-version in ThisBuild := "0.35.0"
+version in ThisBuild := "0.36.0"
